Crate hausarbeit

Ultimate Tic Tac Toe implementation, game agents, heuristics and benchmarking tools

This library provides a full implementation of the Ultimate Tic Tac Toe game. The game can be played by two agents.

Provided agents:

HumanAgent: An agent that requires user input to play.

MiniMaxAgent: An agent that uses the minimax algorithm to determine the best move.

MonteCarloTreeAgent: An agent that uses the Monte Carlo Tree Search algorithm to determine the best move.

RandomAgent: An agent that plays random moves.

Utility agents:

BenchedAgent: An agent that logs the time it takes to make a move.

RandomStartAgent: An agent that uses a random agent for the first depth turns, then switches to another agent.

A custom agent can be implemented by implementing the Agent trait.

Provided heuristics:

CustomHeuristic: A heuristic that uses a custom evaluation function.

MonteCarloGameSearchHeuristic: A heuristic that uses Monte Carlo Tree Search to evaluate the best move.

ParameterizedHeuristic: A heuristic that uses a parameterized evaluation function.

A custom heuristic can be implemented by implementing the Heuristic and MiniBoardHeuristic trait.

Genetic algorithm

The library also contains a GeneticAlgorithm as well as various Selection, Mutation, Recombination and Fitness operators to optimize the weights of the ParameterizedHeuristic.

Usage

Initialize a game with two agents and play it:

use hausarbeit::agent::minimax_agent::MiniMaxAgent;
use hausarbeit::agent::random_agent::RandomAgent;
use hausarbeit::heuristic::custom_heuristic::CustomHeuristic;
use hausarbeit::game::Game;
use hausarbeit::game::player::Player;

let agent_one = MiniMaxAgent::new(2, 1, CustomHeuristic::new(Player::One));
let agent_two = RandomAgent::new();

let mut game = Game::new(Box::new(agent_one), Box::new(agent_two));

game.play();

Modules

agent

Contains the Agent trait and implementations

game

Contains the Game struct

genetic_algorithm

Contains the GeneticAlgorithm struct as well as the genetic algorithm operators

heuristic

Contains the Heuristic and MiniBoardHeuristic and implementations

quality_test

runtime_test

Contains code to test the runtime of the agents