Module hausarbeit::agent::human_agent

Contains the HumanAgent struct

The HumanAgent struct represents an Agent that allows a human player to play the game. The human player can input the moves via the console. The board is printed to the console before each move.

If:

a move is invalid, the player is prompted to input a new move.

the player has to play on a specific board, the board is highlighted by a colorful border.

You can start a game with a human player by calling the start_game_with_human function.

Structs

HumanAgent

An Agent that allows a human player to play the game.

Statics

BACKGROUND_COLOR 🔒

HIGHLIGHT_COLOR 🔒

O_COLOR 🔒

X_COLOR 🔒

Functions

human_against_human

start_game_with_human

Starts a game with a human player.