Module hausarbeit::agent::monte_carlo_tree_agent

Contains the MonteCarloTreeAgent struct

The MonteCarloTreeAgent struct represents an agent that uses the Monte Carlo Tree Search algorithm to evaluate the best move.

The agent uses a Tree to store the game states and the statistics of the nodes.

Modules

tree

Contains the Tree and Node structs

Structs

MonteCarloTreeAgent

Struct representing an agent that uses the Monte Carlo Tree Search algorithm to evaluate the best move

NodeInfo 🔒

Struct representing the information of a node in the tree

Stats 🔒

Struct representing the statistics of a node in the tree

Functions

run