Trait hausarbeit::agent::Agent

pub trait Agent: Send + Sync {
    // Required methods
    fn act(
        &mut self,
        board: UltimateBoard,
        player: Player,
        turn: u32
    ) -> Option<u8>;
    fn get_info(&self) -> AgentInfo;
}

Trait representing an agent that can play Ultimate Tic Tac Toe

Required Methods

fn act(&mut self, board: UltimateBoard, player: Player, turn: u32) -> Option<u8>

The act method is called to get the agent’s move.

The agent should return the index of the field to play on. The index is the human index (0-80) over all boards.

The game will panic if None is returned or if the index is out of bounds. In case of a panic, relevant information about the state of the game will be printed to the console.

It is recommended to return None if the agent cannot play or if the agent can not find a move.

Arguments

board - The current state of the board

Returns

The index of the field to play on

fn get_info(&self) -> AgentInfo

Returns the information of the agent

Returns

The information of the agent