Module hausarbeit::game

Contains the Game struct

The Game struct represents a game of Ultimate Tic Tac Toe. The game is played by two agents.

Modules

bitboard

Contains the BitBoard struct

board

Contains the Board and LegalBoardIterator struct

game_result

Contains the GameResult enum

player

Contains the Player enum

ultimate_board

Contains the UltimateBoard struct

Structs

Game

Struct representing a game of Ultimate Tic Tac Toe