Module hausarbeit::game::bitboard

Contains the BitBoard struct

The BitBoard struct represents a bitboard used for storing the state of the board. The bitboard is a 9-bit integer where each bit represents a square on the board.

Two bitboards are used in Board to represent the state.

The BitBoard struct implements the Not, BitOr, BitAnd, BitXor, BitOrAssign, BitAndAssign, BitXorAssign traits.

Structs

BitBoard

Struct representing a bitboard

BitBoardIterator

Iterator over the set squares in a BitBoard