Struct hausarbeit::game::bitboard::BitBoard

pub struct BitBoard(u16);

Struct representing a bitboard

A bitboard is a 9-bit integer where each bit represents a square on the board.

Fields

0 - The bitboard value as an u16

Tuple Fields
0: u16

9-bit integer encoded as an u16, the upper 7 bits are always unset

Implementations

impl BitBoard

pub fn new(n: u16) -> Self

Creates a new BitBoard with the given value

Arguments

n - The value of the bitboard, must be in 0..2^9

pub const EMPTY: Self = _

Contains a BitBoard with no squares set

pub fn first_square(&self) -> Option<u8>

Returns the index of the first set square in the board. If no square is set, returns None.

Returns

The index of the first set square in the board

pub fn pop_first_square(&mut self) -> Option<u8>

Pops the first set square from the board. If no square is set, returns None.

Returns

The index of the first set square in the board