Module hausarbeit::game::ultimate_board

Contains the UltimateBoard struct

The UltimateBoard struct represents the board of the Ultimate Tic Tac Toe game. The board is represented as a 3x3 array of Board structs.

The UltimateBoard also contains the status of the game and the status of each board as well as the next board to play on.

The board contains a Zobrist hash used to store the evaluation of a Heuristic in a Transposition table.

Structs

UltimateBoard

Struct representing the ultimate board

Enums

BoardIterator

Enum representing the possible iterators for the board

Constants

CENTER_INDEX

The index of the center of a UltimateBoard

CORNER_INDICES

The indices of the corners of a UltimateBoard

EDGE_INDICES

The indices of the edges of a UltimateBoard

NUM_POSITIONS 🔒

Number of squares in Ultimate Tic Tac Toe

NUM_ZOBRIST_VALUES 🔒

Number of Zobrist values needed for the ultimate board

PARTIAL_WIN_POSITIONS 🔒

All possible partial win positions for the ultimate board

WIN_POSITIONS 🔒

All possible win positions for the ultimate board

ZOBRIST_VALUES_NEXT_BOARD_INDEX_OFFSET

Offset of the next_board_index hashes in Zobrist values

Statics

ZOBRIST_VALUES 🔒

Values used for Zobrist hashing