Trait hausarbeit::genetic_algorithm::fitness::FitnessFunction

pub trait FitnessFunction {
    // Required method
    fn calculate_fitness(&self, genes: Vec<Gene>) -> Vec<(Gene, f64)>;

    // Provided method
    fn play_game_with(
        &self,
        lhs: Gene,
        rhs: Gene,
        depth: u32,
        quiescence_depth: u32
    ) -> GameResult { ... }
}

Trait representing a fitness function

The fitness function is used to calculate the fitness of the genes.

The fitness is a value that represents how good the genes are. The higher the fitness, the better the genes.

The fitness can be calculated in different ways.

The Fitness function may be multithreaded.

Required Methods

fn calculate_fitness(&self, genes: Vec<Gene>) -> Vec<(Gene, f64)>

Calculates the fitness of the given genes

Arguments

genes - The genes to calculate the fitness for

Returns

The genes with their fitness

Provided Methods

fn play_game_with( &self, lhs: Gene, rhs: Gene, depth: u32, quiescence_depth: u32 ) -> GameResult

Plays a game with the given genes

Arguments

lhs - The first gene

rhs - The second gene

Returns

The result of the game