Trait hausarbeit::genetic_algorithm::mutation::Mutation

pub trait Mutation {
    // Required method
    fn mutate(&mut self, gene: Gene) -> Gene;

    // Provided method
    fn mutate_all(&mut self, genes: Vec<Gene>) -> Vec<Gene> { ... }
}

Trait representing a mutation

A mutation is a function that changes a gene. The way the gene is changed is determined by the implementation.

Required Methods

fn mutate(&mut self, gene: Gene) -> Gene

Mutates the given gene

Arguments

gene - The gene to mutate

Returns

The mutated gene

Provided Methods

fn mutate_all(&mut self, genes: Vec<Gene>) -> Vec<Gene>

Mutates all the given genes

This function calls Self::mutate for each gene in the list.

Arguments

genes - The genes to mutate

Returns

The mutated genes