Trait hausarbeit::genetic_algorithm::recombination::Recombination

pub trait Recombination {
    // Required method
    fn recombine(&self, lhs: Gene, rhs: Gene) -> (Gene, Gene);

    // Provided method
    fn recombine_all(&self, genes: Vec<Gene>) -> Vec<Gene> { ... }
}

Trait representing a recombination

A recombination is used to recombine two genes. The recombination can be used to create new genes from two given genes.

Required Methods

fn recombine(&self, lhs: Gene, rhs: Gene) -> (Gene, Gene)

Recombines the given genes

Arguments

lhs - The first gene

rhs - The second gene

Returns

The recombined genes

Provided Methods

fn recombine_all(&self, genes: Vec<Gene>) -> Vec<Gene>

Recombines the given genes

Two genes are recombined to create two new genes.

If the amount of genes is odd, then the last gene will be copied without modification.

Arguments

genes - The genes to recombine

Returns

The recombined genes