Module hausarbeit::heuristic

Contains the Heuristic and MiniBoardHeuristic and implementations

The Heuristic trait represents a heuristic that evaluates a UltimateBoard state. The heuristic should return a value that represents how good the board state is.

The MiniBoardHeuristic trait represents a heuristic that evaluates a Board state. The heuristic should return a value that represents how good the board state is.

Modules

custom_heuristic

Contains the CustomHeuristic and CustomMiniBoardHeuristic struct

monte_carlo_game_search_heuristic

Contains the MonteCarloGameSearchHeuristic struct

parameterized_heuristic

Contains the ParameterizedHeuristic and ParameterizedMiniBoardHeuristic struct

Constants

NUM_SMALL_BOARD_STATES

The number of possible legal small board states

Statics

MAX_VALUE

The maximum value a heuristic can return

MIN_VALUE

The minimum value a heuristic can return

Traits

Heuristic

Trait representing a heuristic

MiniBoardHeuristic

Trait representing a heuristic for small boards