Trait hausarbeit::heuristic::Heuristic

pub trait Heuristic: Clone + Send + Sync {
    // Required methods
    fn evaluate(&self, board: UltimateBoard) -> f64;
    fn get_name(&self) -> String;
}

Trait representing a heuristic

A heuristic is a function that evaluates a UltimateBoard state. The heuristic should return a value that represents how good the board state is for a specified player. The maximum and minimum values are defined by the constants MIN_VALUE and MAX_VALUE. The heuristic should return a value between these two values.

Implementations may provide customization options.

Required Methods

fn evaluate(&self, board: UltimateBoard) -> f64

Evaluate the given board state

The heuristic should return a value that represents how good the board state is.

Arguments

board - The board state to evaluate

Returns

The value of the board state, always between MIN_VALUE and MAX_VALUE

fn get_name(&self) -> String

Object Safety

This trait is not object safe.